stack opreation function was redefined to take in an armstate as input

+++ inc/armstate.h
@@ -10,8 +10,8 @@ struct  armstate_t
     memory_t memory;
 
     //Interrupt enable variable
-    static short int RIT;
-    static short int EINT0;
+    static uint8_t RIT;
+    static uint8_t EINT0;
 
     armstate_t(const int mem_endianness);
     armstate_t(void);
@@ -21,5 +21,6 @@ struct  armstate_t
     friend bool operator!=(armstate_t& armstate_lhs, armstate_t& armstate_rhs);
     friend bool operator<(armstate_t& armstate_lhs, armstate_t& armstate_rhs);
 };
+
 
 void print_armstate_diff(armstate_t& armstate_lhs, armstate_t& armstate_rhs, std::ostream& os);
